requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
gunicorn==19.9.0
idna==2.8
influxdb==5.2.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
